diffusion_rs_common
Module core

Re-exports

pub use cpu_backend::CpuStorage;

pub use cpu_backend::CpuStorageRef;

pub use error::Context;

pub use error::Error;

pub use error::Result;

pub use layout::Layout;

pub use shape::Shape;

pub use shape::D;

pub use streaming::StreamTensor;

pub use streaming::StreamingBinOp;

pub use streaming::StreamingModule;

pub use dummy_cuda_backend as cuda;

pub use cuda::CudaDevice;

pub use cuda::CudaStorage;

Modules

backend

backprop

conv

cpu

cpu_backend

display

dummy_cuda_backend

error

layout

npy

Numpy support for tensors.

op

pickle

quantized

safetensors

scalar

shape

The shape of a tensor is a tuple with the size of each of its dimensions.

streaming

test_utils

utils

Structs

DTypeParseError

MetalDevice

MetalStorage

StridedIndex

An iterator over offset position for items of an N-dimensional arrays stored in a flat buffer using some potential strides.

Tensor

The core struct for manipulating tensors.

TensorId

Unique identifier for tensors.

Var

A variable is a wrapper around a tensor, however variables can have their content modified whereas tensors are immutable.

Enums

DType

The different types of elements allowed in tensors.

Device

DeviceLocation

A DeviceLocation represents a physical device whereas multiple Device can live on the same location (typically for cuda devices).

MetalError

Storage

StridedBlocks

TensorIndexer

Generic structure used to index a slice of the tensor

Traits

CustomOp1

Unary ops that can be defined in user-land.

CustomOp2

CustomOp3

FloatDType

IndexOp

Trait used to implement multiple signatures for ease of use of the slicing of a tensor

InplaceOp1

Unary ops that can be defined in user-land. These ops work in place and as such back-prop is unsupported.

InplaceOp2

InplaceOp3

IntDType

Module

ModuleT

NdArray

ToUsize2

WithDType

Functions

from_storage_no_op

Creates a fresh tensor structure based on a storage and a shape, this uses contiguous strides. This has a BackpropOp:none().