diffusion_rs_common::nn::var_builder
Trait Backend

pub trait Backend: Send + Sync {
    type Hints: Default;

    // Required methods
    fn get(
        &self,
        s: Shape,
        name: &str,
        h: Self::Hints,
        dtype: DType,
        dev: &Device,
    ) -> Result<Tensor>;
    fn get_unchecked(
        &self,
        name: &str,
        dtype: DType,
        dev: &Device,
    ) -> Result<Tensor>;
    fn contains_tensor(&self, name: &str) -> bool;
}

A trait that defines how tensor data is retrieved.

Typically this would use disk storage in some specific format, or random initialization. Note that there is a specialized version of this trait (SimpleBackend) that can be used most of the time. The main restriction is that it doesn’t allow for specific args (besides initialization hints).

Required Associated Types

type Hints: Default

Required Methods

fn get( &self, s: Shape, name: &str, h: Self::Hints, dtype: DType, dev: &Device, ) -> Result<Tensor>

Retrieve a tensor with some target shape.

fn get_unchecked( &self, name: &str, dtype: DType, dev: &Device, ) -> Result<Tensor>

Retrieve a tensor based on the name.

fn contains_tensor(&self, name: &str) -> bool