diffusion_rs_common::nn
Module var_builder

A VarBuilder is used to retrieve variables used by a model. These variables can either come from a pre-trained checkpoint, e.g. using VarBuilder::from_mmaped_safetensors, or initialized for training, e.g. using VarBuilder::from_varmap.

Structs

Rename

SafeTensorWithRouting

Shard

ShardedSafeTensors

VarBuilderArgs

A structure used to retrieve variables, these variables can either come from storage or be generated via some form of initialization.

Traits

Backend

A trait that defines how tensor data is retrieved.

Renamer

This traits specifies a way to rename the queried names into names that are stored in an inner VarBuilder.

SimpleBackend

Type Aliases

ShardedVarBuilder

VarBuilder

A simple VarBuilder, this is less generic than VarBuilderArgs but should cover most common use cases.